+++ CHANGELOG.md
@@ -22,7 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- envs: Acceleration filter in `UpkieGroundVelocity` (use a gymnasium.Wrapper)
+- envs: Acceleration filter from `UpkieGroundVelocity` (=> gymnasium.Wrapper)
+- envs: Low-pass filter from `UpkieGroundVelocity` (=> gymnasium.Wrapper)
 
 ## [2.0.0] - 2023-10-26